backend/requirements.txt
fastapi
uvicorn
python-multipart
ultralytics
pillow
opencv-python-headless
pydantic